ファイルアクセスするモジュールを切り出し
